i fix the bot, again
