+++ cartoonfilter_ds_project/handlers.py
@@ -1,7 +1,23 @@
 import os
 #я не знаю что из этого нужно в этом модуле, а что нет, поэтому беру все
 from telegram.ext import Updater, CommandHandler, MessageHandler
-from utilitis import main_keyboard
+from telegram import ReplyKeyboardMarkup
+
+def main_keyboard():
+    return ReplyKeyboardMarkup([
+        ['Neural network'],['Cartoon filter']
+    ])
+
+def remembering_user_choose(update, context):
+    text = update.message.text
+    if text == 'Neural network':
+        context.user_data['filter_type'] = 'Neural network'
+        print(context.user_data['filter_type'])
+   
+    elif text == 'Cartoon filter':
+         context.user_data['filter_type'] = 'Cartoon filter'
+         print(context.user_data['filter_type'])
+
 
 #Приветствуем пользователя
 def greet_user(update, context):
